smart_contracts/requirements.txt
agentics-python
python-dotenv
google-generativeai
crewai
crewai-tools
pyyaml
py-solc-x
web3
eth-tester
py-evm
streamlit-elements